requirements.txt
# 网络安全防护系统依赖包

# Web框架
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1

# 数据库
SQLAlchemy==2.0.21
PyMySQL==1.1.0
mysql-connector-python==8.1.0

# 网络和HTTP
requests==2.31.0
urllib3==2.0.4
dnspython==2.4.2

# 邮件处理
email-validator==2.0.0

# 文件处理
python-magic==0.4.27

# 机器学习
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3
joblib==1.3.2

# 数据处理
beautifulsoup4==4.12.2
lxml==4.9.3

# 任务调度
schedule==1.2.0
APScheduler==3.10.4

# 系统监控
psutil==5.9.5

# 加密和安全
cryptography==41.0.4
bcrypt==4.0.1

# 配置管理
PyYAML==6.0.1

# 时间处理
pytz==2023.3

# 网络安全
python-nmap==0.7.1

# 图像处理
Pillow==10.0.0

# 测试
pytest==7.4.2

# 代码质量
flake8==6.0.0
black==23.7.0